+++ modules/os2forms_nemid/src/Plugin/WebformElement/NemidLastName.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Plugin\WebformElement;
+
+use Drupal\os2web_datalookup\LookupResult\CprLookupResult;
+
+/**
+ * Provides a 'os2forms_nemid_last_name' element.
+ *
+ * @WebformElement(
+ *   id = "os2forms_nemid_last_name",
+ *   label = @Translation("NemID Last name"),
+ *   description = @Translation("Provides a NemID Last name element."),
+ *   category = @Translation("NemID"),
+ * )
+ *
+ * @see \Drupal\os2forms_nemid\Plugin\NemidElementBase
+ * @see \Drupal\os2forms_nemid\Element\NemidLastName
+ */
+class NemidLastName extends ServiceplatformenCprElementBase implements NemidElementPersonalInterface {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getPrepopulateFieldFieldKey(array &$element) {
+    return CprLookupResult::LAST_NAME;
+  }
+
+}