+++ CHANGELOG.md
@@ -11,7 +11,8 @@ before starting to add changes. Use example [placed in the end of the page](#exa
 
 ## [Unreleased]
 
-- Updating ckeditor -> ckeditor5.
+- [PR-326](https://github.com/OS2Forms/os2forms/pull/326)
+  Updating ckeditor -> ckeditor5.
 - [PR-322](https://github.com/OS2Forms/os2forms/pull/322)
   Update Digital Post handler error messages.
 - [PR-320](https://github.com/OS2Forms/os2forms/pull/320)