+++ CHANGELOG.md
@@ -15,6 +15,8 @@ before starting to add changes. Use example [placed in the end of the page](#exa
   signature validation text position) from the *OS2Forms Attachment* element
   onto the *Digital Signature* webform handler. Includes an automatic
   migration update hook for existing webforms.
+- [PR-326](https://github.com/OS2Forms/os2forms/pull/326)
+  Updating ckeditor -> ckeditor5.
 - [PR-322](https://github.com/OS2Forms/os2forms/pull/322)
   Update Digital Post handler error messages.
 - [PR-320](https://github.com/OS2Forms/os2forms/pull/320)